# [SR-3574] optimizing some object columnp agg-function performance and b…