--- a/be/src/exprs/agg/bitmap_intersect.h
+++ b/be/src/exprs/agg/bitmap_intersect.h
@@ -41,9 +41,9 @@ public:
     }
 
     void serialize_to_column(FunctionContext* ctx, ConstAggDataPtr state, Column* to) const override {
-        BitmapValue bitmap = this->data(state).bitmap;
         BitmapColumn* col = down_cast<BitmapColumn*>(to);
-        col->append(&bitmap);
+        BitmapValue& bitmap = const_cast<BitmapValue&>(this->data(state).bitmap);
+        col->append(std::move(bitmap));
     }
 
     void convert_to_serialize_format(const Columns& src, size_t chunk_size, ColumnPtr* dst) const override {
@@ -51,9 +51,9 @@ public:
     }
 
     void finalize_to_column(FunctionContext* ctx, ConstAggDataPtr state, Column* to) const override {
-        BitmapValue bitmap = this->data(state).bitmap;
+        BitmapValue& bitmap = const_cast<BitmapValue&>(this->data(state).bitmap);
         BitmapColumn* col = down_cast<BitmapColumn*>(to);
-        col->append(&bitmap);
+        col->append(std::move(bitmap));
     }
 
     std::string get_name() const override { return "bitmap_intersect"; }