--- a/be/src/exprs/agg/hll_union_count.h
+++ b/be/src/exprs/agg/hll_union_count.h
@@ -76,8 +76,8 @@ public:
                              Column* to) const override {
         DCHECK(to->is_object());
         auto* column = down_cast<HyperLogLogColumn*>(to);
-
-        column->append(&this->data(state));
+        auto& hll_value = const_cast<HyperLogLog&>(this->data(state));
+        column->append(std::move(hll_value));
     }
 
     void convert_to_serialize_format(const Columns& src, size_t chunk_size, ColumnPtr* dst) const override {