…itmap serializing

TEST-1

```
set new_planner_agg_stage=2;

before:
mysql> select bitmap_union_count(v) from bitmap_64;
ERROR 1064 (HY000): The input size for compression should be less than 2113929216

after:
mysql> select bitmap_union_count(v) from bitmap_64;
+-------------------------+
| bitmap_union_count(`v`) |
+-------------------------+
|               100000000 |
+-------------------------+
1 row in set (3 min 35.40 sec)

- UncompressedBytes: 1.21 GB
```

TEST-2

```
set new_planner_agg_stage=2;
mysql> select bitmap_union_count(v) from bitmap_64 < 6;
before Profile:
        DataStreamSender (dst_id=3, dst_fragments=[48252baf120311ec-b3fd00163e04d4c4]):(Active: 40s231ms[40231686641ns], % non-child: 37.30%)
           - PartType: UNPARTITIONED
           - BytesSent: 587.38 MB
           - CompressTime: 2s851ms
           - IgnoreRows: 0
           - OverallThroughput: 14.599884033203125 MB/sec
           - PeakMemoryUsage: 8.00 KB
           - SendRequestTime: 127.595ms
           - SerializeBatchTime: 12s998ms
           - ShuffleDispatchTime: 0ns
           - ShuffleHashTime: 0ns
           - UncompressedBytes: 1.23 GB
           - WaitResponseTime: 23s901ms

after Profile:
        DataStreamSender (dst_id=3, dst_fragments=[3c98f20f120711ec-b3fd00163e04d4c4]):(Active: 39s701ms[39701869028ns], % non-child: 37.20%)
           - PartType: UNPARTITIONED
           - BytesSent: 465.94 MB
           - CompressTime: 1s604ms
           - IgnoreRows: 0
           - OverallThroughput: 11.735963821411133 MB/sec
           - PeakMemoryUsage: 8.00 KB
           - SendRequestTime: 84.157ms
           - SerializeBatchTime: 11s991ms
           - ShuffleDispatchTime: 0ns
           - ShuffleHashTime: 0ns
           - UncompressedBytes: 743.87 MB
           - WaitResponseTime: 25s737ms
```

Agg Function Performance:

```
before:
 AGGREGATION_NODE (id=2):(Active: 1m3s[63909309092ns], % non-child: 33.81%)
           - AggregateFunctions: bitmap_union_count(2: v)
           - AggComputeTime: 25s21ms
           - ExprComputeTime: 972ns
           - ExprReleaseTime: 797ns
           - GetResultsTime: 11s65ms #

AGGREGATION_NODE (id=2):(Active: 56s378ms[56378032377ns], % non-child: 27.49%)
           - AggregateFunctions: bitmap_union_count(2: v)
           - AggComputeTime: 26s763ms
           - ExprComputeTime: 409ns
           - ExprReleaseTime: 463ns
           - GetResultsTime: 35.125us
```