@Seaven 2021-09-11
run starrocks_be_unittest