@stdpain 2021-09-11
I will solve the problem of grayscale upgrade later